Work in progress (Updated 12/16/19)
Main game logic is tested and working.  UI elements tested and working.  Need to a add a restart reset button that kills the current game and starts a new one.  Also the UI needs to be properly styled and modified to look good.  Also need to add card icons to the UI.